Replace Topmenu block with generic Template block
The Topmenu block was used so the topmenu ESI response contains all the
cache tags.

However, this did not work because the Topmenu collects the cache tags
while the menu data is fetched by the phtml template..
The topmenu.phtml template fetches the menu data from the view model, so
the block identities still are empty after the page is rendered.

That said, even if it would work, is no longer is necessary once the
Hyva_Theme module PR is merged to allow view models to implement the
IdentityInterface.
https://gitlab.hyva.io/hyva-themes/magento2-theme-module/-/merge_requests/62

Blocks that should be cached in both the block_html cache and the
Varnish ESI cache need to set the cache_lifetime in layout XML as a
block argument, and also set the cache tags from their view model(s) as
the cache tags like
$block->setData('cache_tags', $viewModel->getIdentities());

At the time of writing this is only used for the top navigation block.